+++ src/Gnugat/Phpixture/ValueSwitcher/is_to_one_relationship.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Gnugat\Phpixture\ValueSwitcher;
+
+/**
+ * @param mixed $value
+ *
+ * @return bool
+ */
+function is_to_one_relationship($value) {
+    return is_string($value) && !empty($value) && '@' === $value[0];
+}